requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-socketio==5.10.0
python-engineio==4.9.0
eventlet==0.33.3
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

